+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # Go-Utils
 
+## v0.1.7
+
+**2021-05-21**
+
+- 增加 `MustJSONIndent` `MustJSONIndentString`
+- 启用可选 JSON 库: `goccy/go-json@v0.5.1`, 编译参数: `go build -tags=gojson .`
+- 升级 `json-iterator/go@v1.1.11`, 默认使用: `ConfigCompatibleWithStandardLibrary`
+
 ## v0.1.6
 
 **2021-05-20**
@@ -53,8 +61,8 @@
 - 加密小工具加密方式改为: `AesCBCEnStringB58(value, MD5Hex(secret))`
 - `json` 使用 `gin` 类似的可选组织方式:
   - `go build .` 默认使用 `json-iterator/go`
-  - `go build -tags=gojson.` 使用标准 JSON 库 `encoding/json`
-  - `go build -tags=go_json .` 使用 `goccy/go-json` (暂不成熟, 观望中)
+  - `go build -tags=stdjson.` 使用标准 JSON 库 `encoding/json`
+  - `go build -tags=gojson .` 使用 `goccy/go-json`
 
 ## v0.0.2